+++ packages/xhr-mock/src/MockHeaders.ts
@@ -1,1 +1,0 @@
-export type MockHeaders = {[name: string]: string};