+++ packages/xhr-mock/src/router/XHRMockError.ts
@@ -1,5 +1,0 @@
-export default class XHRMockError extends Error {
-  constructor(message?: string) {
-    super(`xhr-mock: ${message}`);
-  }
-}